pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deprecated>=1.2.18",
    "fastmcp>=2.12.5",
    "gcloud>=0.18.3",
    "google-adk[database]==0.3.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "ir>=0.0.6",
    "litellm==1.66.3",
    "psutil==5.9.5",
    "python-dotenv==1.1.0",
    "stdioconnectionparams",
    "yfinance==0.2.56",
]
